Address PR review feedback
- Fix hasEvent null check (undefined vs null)
- Fix UnixTimestamp to allow 0 (nonNegative)
- Extract shared FilterMatcher module (DRY)
- Move connection counter inside make (no global state)
